+++ fugitive-darkness-api/src/main/java/com/zer0s2m/fugitivedarkness/api/handlers/ControllerApiGitRepoDelete.java
@@ -2,6 +2,7 @@ package com.zer0s2m.fugitivedarkness.api.handlers;
 
 import com.zer0s2m.fugitivedarkness.api.exception.NotFoundException;
 import com.zer0s2m.fugitivedarkness.common.dto.ContainerGitRepoControl;
+import com.zer0s2m.fugitivedarkness.models.GitRepoModel;
 import com.zer0s2m.fugitivedarkness.provider.GitRepo;
 import com.zer0s2m.fugitivedarkness.provider.HelperGitRepo;
 import com.zer0s2m.fugitivedarkness.repository.GitRepoRepository;
@@ -14,6 +15,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.file.Path;
+import java.util.List;
 
 /**
  * Request handler to remove a git repository from the system.
@@ -31,43 +33,70 @@ final public class ControllerApiGitRepoDelete implements Handler<RoutingContext>
      */
     @Override
     public void handle(@NotNull RoutingContext event) {
-        ContainerGitRepoControl containerGitRepoDelete = event
+        final GitRepoRepository repositoryGit = new GitRepoRepositoryImpl(event.vertx());
+        final ContainerGitRepoControl containerGitRepoDelete = event
                 .body()
                 .asJsonObject()
                 .mapTo(ContainerGitRepoControl.class);
+
         boolean isExistsGitRepository = HelperGitRepo.existsGitRepository(
                 containerGitRepoDelete.group(), containerGitRepoDelete.project());
 
-        if (!isExistsGitRepository) {
-            event.fail(
-                    HttpResponseStatus.NOT_FOUND.code(),
-                    new NotFoundException("Object not found in the system"));
-        } else {
-            event.vertx()
-                    .executeBlocking(() -> {
-                        final Path sourceGitRepository = HelperGitRepo.getSourceGitRepository(
-                                containerGitRepoDelete.group(), containerGitRepoDelete.project());
-                        logger.info("Start deleting a git repository [" + sourceGitRepository + "]");
-                        gitRepo.gDelete(containerGitRepoDelete.group(), containerGitRepoDelete.project());
-                        logger.info("Finish deleting git repository [" + sourceGitRepository + "]");
-                        return null;
-                    }, false)
-                    .onSuccess(handler -> {
-                        final GitRepoRepository repositoryGit = new GitRepoRepositoryImpl(event.vertx());
-                        repositoryGit
-                                .deleteByGroupAndProject(
-                                        containerGitRepoDelete.group(), containerGitRepoDelete.project())
-                                .onFailure(fail -> {
-                                    logger.error(" Failure (DB): " + fail.fillInStackTrace());
-                                    repositoryGit.closeClient();
-                                })
-                                .onSuccess((ar) -> repositoryGit.closeClient());
-                    });
-            event
-                    .response()
-                    .setStatusCode(HttpResponseStatus.NO_CONTENT.code());
-            event.next();
-        }
+        repositoryGit.findByGroupAndProject(
+                        containerGitRepoDelete.group(),
+                        containerGitRepoDelete.project())
+                .onSuccess(ar1 -> {
+                    final List<GitRepoModel> getRepositories = repositoryGit.mapTo(ar1);
+                    boolean isLocalRepo;
+                    if (!getRepositories.isEmpty()) {
+                        isLocalRepo = getRepositories.get(0).getIsLocal();
+                    } else {
+                        isLocalRepo = false;
+                    }
+
+                    if (!isExistsGitRepository && !isLocalRepo) {
+                        event.fail(
+                                HttpResponseStatus.NOT_FOUND.code(),
+                                new NotFoundException("Object not found in the system"));
+                    } else {
+                        event.vertx()
+                                .executeBlocking(() -> {
+                                    if (!isLocalRepo) {
+                                        final Path sourceGitRepository = HelperGitRepo.getSourceGitRepository(
+                                                containerGitRepoDelete.group(), containerGitRepoDelete.project());
+                                        logger.info("Start deleting a git repository [" + sourceGitRepository + "]");
+                                        gitRepo.gDelete(containerGitRepoDelete.group(), containerGitRepoDelete.project());
+                                        logger.info("Finish deleting git repository [" + sourceGitRepository + "]");
+                                    }
+
+                                    return null;
+                                }, false)
+                                .onSuccess(handler -> {
+                                    repositoryGit
+                                            .deleteByGroupAndProject(
+                                                    containerGitRepoDelete.group(), containerGitRepoDelete.project())
+                                            .onFailure(fail -> {
+                                                logger.error(" Failure (DB): " + fail.fillInStackTrace());
+                                                repositoryGit.closeClient();
+                                            })
+                                            .onSuccess((ar2) -> repositoryGit.closeClient());
+
+                                    event
+                                            .response()
+                                            .setStatusCode(HttpResponseStatus.NO_CONTENT.code());
+                                    event.next();
+                                });
+                    }
+                })
+                .onFailure(fail -> {
+                    logger.error(" Failure (DB): " + fail.fillInStackTrace());
+                    repositoryGit.closeClient();
+
+                    event
+                            .response()
+                            .setStatusCode(HttpResponseStatus.INTERNAL_SERVER_ERROR.code());
+                    event.next();
+                });
     }
 
 }