+++ fugitive-darkness-common/src/main/java/com/zer0s2m/fugitivedarkness/common/dto/ContainerGitRepoInstall.java
@@ -2,7 +2,9 @@ package com.zer0s2m.fugitivedarkness.common.dto;
 
 public record ContainerGitRepoInstall(
 
-        String remote
+        String remote,
+
+        boolean isLocal
 
 ) {
 }